Move editor's file to the library in order to be able to test
