+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/BufferValidationHelpers.cpp
