+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/BufferValidationHelpers.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Helpers.h>
+#include <keyboardmanager/common/Helpers.h>
 
 namespace BufferValidationHelpers
 {
