+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/Constants.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "pch.h"
+
+namespace KeyboardManagerConstants
+{
+    // String constant for the default app name in Remap shortcuts
+    inline const std::wstring DefaultAppName = GET_RESOURCE_STRING(IDS_EDITSHORTCUTS_ALLAPPS);
+}
