+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/Dialog.cpp
