+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/EditShortcutsWindow.cpp
