+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/EditShortcutsWindow.h
