+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/ShortcutControl.cpp
