+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/ShortcutControl.h
