+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/XamlBridge.cpp
