+++ src/modules/keyboardmanager/test/BufferValidationTests.cpp
@@ -1,10 +1,11 @@
 #include "pch.h"
 #include "CppUnitTest.h"
-#include <keyboardmanager/KeyboardManagerEditor/BufferValidationHelpers.h>
+#include <keyboardmanager/KeyboardManagerEditorLibrary/BufferValidationHelpers.h>
 #include "TestHelpers.h"
 #include <common/interop/keyboard_layout.h>
 #include <common/interop/shared_constants.h>
 #include <functional>
+#include <keyboardmanager/common/ErrorTypes.h>
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 using namespace TestHelpers;
