+++ src/modules/keyboardmanager/test/LoadingAndSavingRemappingTests.cpp
@@ -1,9 +1,10 @@
 #include "pch.h"
 #include "CppUnitTest.h"
 #include <keyboardmanager/common/KeyboardManagerState.h>
-#include <keyboardmanager/KeyboardManagerEditor/LoadingAndSavingRemappingHelper.h>
+#include <keyboardmanager/KeyboardManagerEditorLibrary/LoadingAndSavingRemappingHelper.h>
 #include "TestHelpers.h"
 #include <common/interop/shared_constants.h>
+#include <keyboardmanager/common/ErrorTypes.h>
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
